requirements.txt
biopython==1.79
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.2
Pillow==8.4.0
psycopg2==2.9.2
pymprog==1.1.2
pyprog==1.1.0.post2
python-dotenv==0.19.2
requests==2.26.0
style==1.1.0
swiglpk==5.0.3
update==0.0.1
urllib3==1.26.6
Werkzeug==2.0.2
matplotlib==3.5.0